# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ForgeRun Labs
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  LIVE: "LIVE",
  BUILDING: "BUILDING",
  OFFLINE: "OFFLINE",
} as const;

export interface Project {
  id: number;
  name: string;
  description: string;
  slug: string;
  url: string;
  status: ProjectStatus;
  isPublic: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface CreateProjectInput {
  name: string;
  description: string;
  url: string;
  status?: ProjectStatus;
  isPublic?: boolean;
  sortOrder?: number;
}

export interface UpdateProjectInput {
  name?: string;
  description?: string;
  url?: string;
  status?: ProjectStatus;
  isPublic?: boolean;
  sortOrder?: number;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  status: string;
}