# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for ForgeRun Labs
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project registry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listPublicProjects
      tags: [projects]
      summary: List public projects
      description: Returns all public projects for the public homepage
      responses:
        "200":
          description: List of public projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /admin/projects:
    get:
      operationId: listAllProjects
      tags: [projects]
      summary: List all projects (admin)
      description: Returns all projects including private ones
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of all projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin)
      parameters:
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/projects/{id}:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ProjectStatus:
      type: string
      enum:
        - LIVE
        - BUILDING
        - OFFLINE

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        slug:
          type: string
        url:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        isPublic:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - slug
        - url
        - status
        - isPublic
        - sortOrder
        - createdAt

    CreateProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        isPublic:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - description
        - url

    UpdateProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        isPublic:
          type: boolean
        sortOrder:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResponse:
      type: object
      properties:
        status:
          type: string
      required:
        - status