# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ForgeRun Labs
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from "./projectStatus";

export interface CreateProjectInput {
  name: string;
  description: string;
  url: string;
  status?: ProjectStatus;
  isPublic?: boolean;
  sortOrder?: number;
}