# lib/api-zod/src/generated/types/projectStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ForgeRun Labs
 * OpenAPI spec version: 0.1.0
 */

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  LIVE: "LIVE",
  BUILDING: "BUILDING",
  OFFLINE: "OFFLINE",
} as const;